* commit '31f909667ab9aa47033f0787d0627bf51433ab8d':
  one scan can has more than one process
  change ids to long
  change ids to long
  bump minor version
  increase numeric data types
  change line of code type to long

+++ api/src/main/java/com/ripstech/apiconnector2/ApiRequest.java
@@ -50,7 +50,7 @@ public class ApiRequest {
 		return this;
 	}
 
-	public ApiRequest appendPath(int path) {
+	public ApiRequest appendPath(long path) {
 		this.path += "/" + path;
 		return this;
 	}
