+++ api/src/main/java/com/ripstech/apiconnector2/entity/receive/application/scan/File.java
@@ -2,16 +2,16 @@ package com.ripstech.apiconnector2.entity.receive.application.scan;
 
 public class File {
 
-	private int id;
-	private Integer loc;
+	private long id;
+	private Long loc;
 	private String path;
 	private String code;
 
-	public int getId() {
+	public long getId() {
 		return this.id;
 	}
 
-	public Integer getLoc() {
+	public Long getLoc() {
 		return this.loc;
 	}
