+++ api/src/main/java/com/ripstech/apiconnector2/entity/receive/application/scan/issue/Type.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public class Type {
 
-	private int id;
+	private long id;
 	private String tag;
 	private String description;
 	private String name;
@@ -26,7 +26,7 @@ public class Type {
 	private List<Resource> resources = Collections.emptyList();
 	private Boolean inactive;
 
-	public int getId() {
+	public long getId() {
 		return this.id;
 	}
 
@@ -104,13 +104,13 @@ public class Type {
 
 	public static class Resource {
 
-		private int id;
+		private long id;
 		private String title;
 		private String author;
 		private String url;
 		private OffsetDateTime publishedAt;
 
-		public int getId() {
+		public long getId() {
 			return id;
 		}
