+++ api/src/main/java/com/ripstech/apiconnector2/service/application/scan/ProcessService.java
@@ -15,10 +15,10 @@ import static com.ripstech.apiconnector2.service.template.GenericService.HttpMet
 
 public class ProcessService extends PostGetService<Process, ProcessPost> {
 
-	private final int applicationId;
-	private final int scanId;
+	private final long applicationId;
+	private final long scanId;
 
-	public ProcessService(String baseUri, int applicationId, int scanId) {
+	public ProcessService(String baseUri, long applicationId, long scanId) {
 		super(baseUri);
 		this.applicationId = applicationId;
 		this.scanId = scanId;
